backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
django-cors-headers==4.3.1
Pillow==10.1.0
djoser==2.2.2
python-decouple==3.8
celery==5.3.6
redis==5.0.1 